TRANSFORMAÇÃO DE DADOS/requirements.txt
numpy==1.26.3
pandas==2.1.1
tabula-py==2.7.0
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2